first factor
